SEO Tweaks
Used feedthebot.com to get feedback on what changes needed to be made.
Setting cache expiration and compressing the html through the .htaccess
file. Changed the robot.txt to tell the crawlers to ignore the search
pages. Switch the home page to using list-group so that the links had
more area for mobile users to touch.